refactor(project structure and trivial things) and reformat
refactor(project structure and trivial things) and reformat

refactor(project structure and trivial things) and reformat
