+++ backend/api/voice_router.py
@@ -3,25 +3,27 @@ from typing import Annotated
 from uuid import uuid4, UUID
 
 from fastapi import APIRouter, Depends, UploadFile, HTTPException, Response
-from mypy_boto3_s3 import S3Client
 
 from auth import JwtAuth
-from core.dependencies import get_voice_repository, get_user_repository
-from db.entity import User
-from db.entity import Voice
-from db.repository import VoiceRepository, UserRepository
-from s3_service import get_s3_client, upload_audio, download_audio
+from models import User, Voice
+from repository import (
+    get_voice_repository,
+    get_user_repository,
+    VoiceRepository,
+    UserRepository,
+)
+from s3_service import upload_audio, download_audio
 
 router = APIRouter(dependencies=[Depends(JwtAuth())])
 
 
 @router.post("/voice")
 async def upload_voice(
-        to_user: int,
-        audio_file: UploadFile,
-        from_user_kakao_id: Annotated[int, Depends(JwtAuth())],
-        voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
-        user_repo: Annotated[UserRepository, Depends(get_user_repository)],
+    to_user: int,
+    audio_file: UploadFile,
+    from_user_kakao_id: Annotated[int, Depends(JwtAuth())],
+    voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
+    user_repo: Annotated[UserRepository, Depends(get_user_repository)],
 ):
     # save voice data to db
     from_user: User = user_repo.find_by_kakao_id(from_user_kakao_id)
@@ -31,8 +33,7 @@ async def upload_voice(
     voice_repo.insert(voice)
 
     # upload to s3
-    s3_client: S3Client = get_s3_client()
-    upload_audio(s3_client, f"{s3_id}.mp3", await audio_file.read())
+    upload_audio(f"{s3_id}.mp3", await audio_file.read())
 
     # (tmp)
     return "ok"
@@ -40,10 +41,10 @@ async def upload_voice(
 
 @router.get("/voice/{voice_id}/meta")
 async def get_voice_metadata(
-        voice_id: int,
-        user_kakao_id: Annotated[int, Depends(JwtAuth())],
-        voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
-        user_repo: Annotated[UserRepository, Depends(get_user_repository)],
+    voice_id: int,
+    user_kakao_id: Annotated[int, Depends(JwtAuth())],
+    voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
+    user_repo: Annotated[UserRepository, Depends(get_user_repository)],
 ):
     # find voice
     voice = voice_repo.find_by_id(voice_id)
@@ -70,37 +71,35 @@ async def get_voice_metadata(
 
 @router.get("/voice/{voice_id}/audio")
 async def get_voice_audio(
-        voice_id: int,
-        user_kakao_id: Annotated[int, Depends(JwtAuth())],
-        user_repo: Annotated[UserRepository, Depends(get_user_repository)],
-        voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
+    voice_id: int,
+    user_kakao_id: Annotated[int, Depends(JwtAuth())],
+    user_repo: Annotated[UserRepository, Depends(get_user_repository)],
+    voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
 ):
     # find voice
     voice = voice_repo.find_by_id(voice_id)
-    if not voice:
+    if voice is None:
         raise HTTPException(status_code=500, detail="voice required")
 
     # authorize
     # authorize: 보낸 쪽(from_user), 받는 쪽(to_user)모두 권한 있임
     user: User = user_repo.find_by_kakao_id(user_kakao_id)
-    assert not user
+    assert user is not None
     if user.id not in [voice.to_user, voice.from_user]:
         raise HTTPException(status_code=401, detail="unauthorized")
 
     s3_id = str(UUID(bytes=voice.s3_id))
-    file_path = pathlib.Path.cwd().absolute() / "audio" / f"{s3_id}.mp3"
 
-    s3_client: S3Client = get_s3_client()
-    audio_binary: bytes = download_audio(s3_client, f"{s3_id}.mp3")
+    audio_binary: bytes = download_audio(f"{s3_id}.mp3")
 
     return Response(content=audio_binary, media_type="audio/mpeg")
 
 
 @router.get("/voice/sent")
 async def get_voice_id_list(
-        user_kakao_id: Annotated[int, Depends(JwtAuth())],
-        user_repo: Annotated[UserRepository, Depends(get_user_repository)],
-        voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
+    user_kakao_id: Annotated[int, Depends(JwtAuth())],
+    user_repo: Annotated[UserRepository, Depends(get_user_repository)],
+    voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
 ) -> list[int]:
     user: User = user_repo.find_by_kakao_id(user_kakao_id)
     assert user is not None
@@ -114,9 +113,9 @@ async def get_voice_id_list(
 
 @router.get("/voice/received")
 async def get_voice_id_list(
-        user_kakao_id: Annotated[int, Depends(JwtAuth())],
-        user_repo: Annotated[UserRepository, Depends(get_user_repository)],
-        voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
+    user_kakao_id: Annotated[int, Depends(JwtAuth())],
+    user_repo: Annotated[UserRepository, Depends(get_user_repository)],
+    voice_repo: Annotated[VoiceRepository, Depends(get_voice_repository)],
 ) -> list[int]:
     user: User = user_repo.find_by_kakao_id(user_kakao_id)
     assert user is not None
