+++ backend/data/schema.sql
@@ -17,14 +17,3 @@ create table voice
     is_correct BOOLEAN DEFAULT FALSE,
     created_at TIMESTAMP
 );
-
-# test data
-INSERT INTO users (kakao_id, username, nickname)
-VALUES (3750899517, 'user1', 'nick1'),
-       (3750899518, 'user2', 'nick2'),
-       (3750899519, 'user3', 'nick3');
-
-INSERT INTO voice (s3_id, from_user, to_user)
-VALUES (UUID_TO_BIN('da06e564-fc83-44f3-a355-6c1f8d3afac0'), 1, 2),
-       (UUID_TO_BIN('da06e564-fc83-44f4-a355-6c1f8d3afac0'), 1, 3),
-       (UUID_TO_BIN('da06e564-fc83-44f5-a355-6c1f8d3afac0'), 2, 1);
