--- backend/db/entity/__init__.py
@@ -1,3 +1,0 @@
-from .entity import User, Voice
-
-__all__ = ["User", "Voice"]
