--- backend/db/repository/__init__.py
@@ -1,7 +1,0 @@
-from .user_repository import UserRepository
-from .voice_repository import VoiceRepository
-
-__all__ = [
-    "VoiceRepository",
-    "UserRepository",
-]
