+++ backend/models/__init__.py
@@ -1,0 +1,4 @@
+from .user import User
+from .voice import Voice
+
+__all__ = ["User", "Voice"]
