+++ backend/repository/__init__.py
@@ -1,0 +1,10 @@
+from .dependencies import get_user_repository, get_voice_repository
+from .voice_repository import VoiceRepository
+from .user_repository import UserRepository
+
+__all__ = [
+    "get_user_repository",
+    "get_voice_repository",
+    "UserRepository",
+    "VoiceRepository",
+]
